# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FO76 DB",
    "description": "Локальный API утилиты FO76 DB: каталог предметов, инвентарь, роллы, события, настройки.\n\n**Доступ.** С этого компьютера (127.0.0.1, без прокси) токен не нужен. Из сети, если в настройках задан токен, передавайте\n`Authorization: Bearer <токен>` (или cookie `fo76db_token`, или `?token=` один раз). Без токена — 401.\nИзменяющие запросы с чужого `Origin` без Bearer отклоняются (защита от CSRF). Для браузерного фронтенда на другом адресе\nдобавьте его в `cors_origins` в `config.toml`. Запись путей, настроек сети, конфига IOM и подобное — только с этого компьютера.",
    "version": "2026.10.5"
  },
  "paths": {
...
    "/api/profiles": {
      "get": {
        "summary": "Api Profiles",
        "operationId": "api_profiles_api_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "tags": [
          "profiles"
        ]
      },
      "post": {
        "summary": "Api Profile Save",
        "operationId": "api_profile_save_api_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_profile_save_api_profiles_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/accounts": {
      "post": {
        "summary": "Api Account Save",
        "operationId": "api_account_save_api_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_account_save_api_accounts_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "accounts"
        ]
      }
    },
...
    "/api/snapshots": {
      "get": {
        "summary": "Api Snapshots",
        "operationId": "api_snapshots_api_snapshots_get",
        "parameters": [
          {
            "name": "character_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Character Id"
            }
          },
          {
            "name": "archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/diff": {
      "get": {
        "summary": "Api Diff",
        "description": "Что изменилось между снимками a (раньше) и b (позже) одного персонажа.",
        "operationId": "api_diff_api_diff_get",
        "parameters": [
          {
            "name": "a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "A"
            }
          },
          {
            "name": "b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "B"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "diff"
        ]
      }
    },
    "/api/snapshots/archive": {
      "post": {
        "summary": "Api Snapshots Archive",
        "description": "Скрыть старые снимки, оставив последние keep (≥1). Данные не удаляются.",
        "operationId": "api_snapshots_archive_api_snapshots_archive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_api_snapshots_archive_api_snapshots_archive_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "snapshots"
        ]
      }
    },
    "/api/snapshots/{sid}/unarchive": {
      "post": {
        "summary": "Api Snapshot Unarchive",
        "operationId": "api_snapshot_unarchive_api_snapshots__sid__unarchive_post",
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "tags": [
          "snapshots"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_api_account_save_api_accounts_post": {
        "properties": {
          "account": {
            "type": "string",
            "title": "Account"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "account"
        ],
        "title": "Body_api_account_save_api_accounts_post"
      },
...
      "Body_api_profile_save_api_profiles_post": {
        "properties": {
          "character_id": {
            "type": "integer",
            "title": "Character Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "mule": {
            "type": "boolean",
            "title": "Mule",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "character_id"
        ],
        "title": "Body_api_profile_save_api_profiles_post"
      },
...
      "Body_api_snapshots_archive_api_snapshots_archive_post": {
        "properties": {
          "character_id": {
            "type": "integer",
            "title": "Character Id"
          },
          "keep": {
            "type": "integer",
            "title": "Keep"
          }
        },
        "type": "object",
        "required": [
          "character_id",
          "keep"
        ],
        "title": "Body_api_snapshots_archive_api_snapshots_archive_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "token из config.toml / «Настройки → Доступ по сети»"
      }
    }
  },
...
}